requirements.txt
environs~=9.5.0
python-telegram-bot~=13.15
redis~=4.5.1
